backend/requirements/base.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-filter==24.3
django-environ==0.11.2
psycopg2-binary==2.9.9
django-storages==1.14.4
boto3==1.35.36
Pillow==10.4.0
